+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+<<<<<<< HEAD
 ### Added
 - Ignore diagnostics for files using `phpCodeSniffer.ignorePatterns` option.
+=======
+### Fixed
+- "Ignore * for this line" action should be present for all diagnostics.
+>>>>>>> f8802f9 (Fixed "Ignore * for this line" Actions)
 
 ## [1.0.0] - 2021-03-01
 ### Fixed